Improve syntax highlighting for urls in paket.dependencies file
Http urls were being highlighted as a keyword and a comment. Now urls
are identified and highlighted as strings.

The addin is now using a custom syntax mode implementation for
paket.dependencies files. This custom syntax mode uses the
HighlightUrlSemanticRule which allows urls to be recognised by
MonoDevelop so you can now right click the url in the text editor and
select Open URL... to open it with the default browser.