Fix comment syntax highlighting.
Remove multiline comments.
Remove comment support from all paket files except paket.dependency.
Comments now only work if a comment string '#' or '//' begins a line
ignoring whitespace. Previously a comment was highlighted anywhere on
a line.

Using a span with flags="FirstNonWs" does not seem to work for
comments. The change log syntax highlighting uses this to highlight a
comment line:

    * comment :

This comment is not highlighted in MonoDevelop. As a workaround the
syntax highlighting file for paket.dependencies is using a regex match
to match the entire line.