Add paket restore.
Using a solution side paket restore instead of having each project
import the paket.targets file.

This uses a Before.MonoDevelop.Paket.sln.targets file which only works
when running a command line build. A modified version of the
paket.targets is used to run the restore for the solution instead of
running a restore for each project's paket.references file.