Support running paket init from the unified search window.
Can now create a paket.dependencies file in the solution's directory
by selecting "paket init" from the unified search window. This runs
paket.exe which is included with the addin.

The status bar is updated as the paket init progresses.

Output from paket.exe is reported in a Paket Console window.