Support installing a NuGet package to a project in unified search.
Runs 'paket add nuget packageId version packageversion project path'
to add a NuGet package to the currently selected project. The project
has to be selected in the Solution window.

The project file will be reloaded by MonoDevelop after the command
has run so the new references will be reflected in the Solution
window.