Allow NuGet packages to be added from the Paket Dependencies folder.
The Paket Dependencies folder now has an Add Packages menu which
displays the same Add Packages dialog used by the NuGet addin.
Selecting a NuGet package and clicking the Add Package button will
add the NuGet package to the paket.dependencies file and install
it into the packages directory.

Imported quite a bit of the NuGet addin source code so the Paket addin
does not take a dependency on the NuGet addin and also allows the
behaviour of the Add Packages dialog to be modified.