Get NuGet package ids in cache using filename.
Using the NuGet's MachineCache will open each of the .nupkg files in
the local machine's cache directory. Getting the package id from the
filename is simpler but the case may not match the package id in some
cases. It also is gets around a bug in Mono where it cannot open
a NuGet package and a null reference exception is thrown:

 Object reference not set to an instance of an object
at System.IO.Packaging.Package.LoadRelationships ()

https://bugzilla.xamarin.com/show_bug.cgi?id=26205