Refresh UI and files after failing to remove last NuGet package
Paket currently throws an error when removing the last NuGet package:

Group Main was not found in paket.dependencies.

https://github.com/fsprojects/Paket/issues/1950

This left the UI and files still showing the original information even
though Paket does remove the NuGet package from the dependencies and
references files. So now the addin still notifies that these files
have changed when there was an error removing a NuGet package. This
ensures the Solution window is updated with the correct information
and the files are reloaded if they are open in the text editor.