Support installing specific versions.
The Add Packages dialog now has "Latest" as a version in the combo
box. If Latest is selected then the NuGet package is added without a
version. If a specific version is selected then the version will be
added to the paket.dependencies file.

For pre-release NuGet packages the version is used even if Latest is
selected otherwise Paket will use the latest stable version.