Fix errors running paket.exe not reported in Paket Console
Seems to be a bug in MonoDevelop's ProcessService.StartConsoleProcess
where the exit code for the command run is not available when it
is set on the ProcessAsyncOperation. To workaround this the code
checks if Console.Error was used and if so reports an error even if
the exit code is zero.