Fix code completion not working in paket.dependencies
Visual Studio for Mac 8.1 needs a document controller extension
and a custom DocumentContext in order for the paket text editor
extension to be used. Also work around a bug in Visual Studio for Mac
that means mime types that need a regex or are an exact match are
not matched with the filename since the full filename is being used
when attempting to match the mime type. This resulted in the
document extension and the document controller not being used since
the mime type 'application/paket-dependencies' is not associated
with the paket.dependencies file.